+++ src/lib/components/forms/form/model/form-type-map.ts
@@ -3,7 +3,6 @@ import { CheckboxGroupComponent } from '../../checkbox-group/checkbox-group.comp
 import { TerraFormTypeInterface } from './terra-form-type.interface';
 import { TerraDatePickerComponent } from '../../input/date-picker/terra-date-picker.component';
 import { TerraFileInputComponent } from '../../input/file-input/terra-file-input.component';
-import { TerraTextInputComponent } from '../../input/text-input/terra-text-input.component';
 import { NumberInputComponent } from '../form-components/number-input/number-input.component';
 import { TerraDoubleInputComponent } from '../../input/double-input/terra-double-input.component';
 import { SelectComponent } from '../form-components/select/select.component';
@@ -14,6 +13,7 @@ import { TerraCodeEditorComponent } from '../../../editors/code-editor/terra-cod
 import { TerraSliderComponent } from '../../slider/terra-slider.component';
 import { TerraPortletComponent } from '../../../layouts/portlet/terra-portlet.component';
 import { TerraSuggestionBoxComponent } from '../../suggestion-box/terra-suggestion-box.component';
+import { TextInputComponent } from '../form-components/text-input/text-input.component';
 import { CheckboxComponent } from '../form-components/checkbox/checkbox.component';
 import { TextAreaComponent } from '../form-components/text-area/text-area.component';
 
@@ -42,9 +42,10 @@ export class FormTypeMap {
     };
     public readonly text: TerraFormTypeInterface = {
         inputMap: {
-            required: 'inputIsRequired'
+            required: 'isRequired',
+            iban: 'isIban'
         },
-        component: TerraTextInputComponent
+        component: TextInputComponent
     };
     public readonly textarea: TerraFormTypeInterface = {
         inputMap: {